Update number of rectangles parameter to take in an int.

+++ Rectangle.DomainLogic/Services/Implementations/GridService.cs
@@ -13,7 +13,7 @@ namespace Rectangle.DomainLogic.Services.Implementations
         {
         }
 
-        public Grid InitialiseWithRectanglesOfRandomSize(byte numberOfRectangles)
+        public Grid InitialiseWithRectanglesOfRandomSize(int numberOfRectangles)
         {
             AssertNumberOfRectanglesInRange(numberOfRectangles);
